requirements.txt
requests>=2.28
pydub>=0.25
openai-whisper>=1.2
